rstudio/enabled

+++ recipes/default.rb
@@ -1424,54 +1424,59 @@ end
 #
 # Rstudio
 #
-case node['platform']
- when 'debian', 'ubuntu'
-   package "r-base"
 
-  remote_file "#{Chef::Config['file_cache_path']}/#{node['rstudio']['deb']}" do
-    user node['glassfish']['user']
-    group node['glassfish']['group']
-    source node['download_url'] + "/#{node['rstudio']['deb']}"
-    mode 0755
-    action :create
-  end
-   
-   bash 'install_rstudio_debian' do
-    user "root"
-    code <<-EOF
+if node['rstudio']['enabled'].eql? "true"
+
+  case node['platform']
+  when 'debian', 'ubuntu'
+    package "r-base"
+
+    remote_file "#{Chef::Config['file_cache_path']}/#{node['rstudio']['deb']}" do
+      user node['glassfish']['user']
+      group node['glassfish']['group']
+      source node['download_url'] + "/#{node['rstudio']['deb']}"
+      mode 0755
+      action :create
+    end
+    
+    bash 'install_rstudio_debian' do
+      user "root"
+      code <<-EOF
       set -e
       cd #{Chef::Config['file_cache_path']}
       apt-get install gdebi-core -y
       gdebi #{node['rstudio']['deb']}
     EOF
-   end
-   
- when 'redhat', 'centos', 'fedora'
+    end
+    
+  when 'redhat', 'centos', 'fedora'
 
-  remote_file "#{Chef::Config['file_cache_path']}/#{node['rstudio']['rpm']}" do
-    user node['glassfish']['user']
-    group node['glassfish']['group']
-    source node['download_url'] + "/#{node['rstudio']['rpm']}"
-    mode 0755
-    action :create
-  end
+    remote_file "#{Chef::Config['file_cache_path']}/#{node['rstudio']['rpm']}" do
+      user node['glassfish']['user']
+      group node['glassfish']['group']
+      source node['download_url'] + "/#{node['rstudio']['rpm']}"
+      mode 0755
+      action :create
+    end
 
-   bash 'install_rstudio_rhel' do
-    user "root"
-    code <<-EOF
+    bash 'install_rstudio_rhel' do
+      user "root"
+      code <<-EOF
       set -e
       cd #{Chef::Config['file_cache_path']}
       yum install --nogpgcheck #{node['rstudio']['rpm']}
     EOF
-   end
-   
-end   
+    end
+    
+  end   
 
-bash 'disable_rstudio_systemd_daemons' do
-  user "root"
-  ignore_failure true
-  code <<-EOF
+  bash 'disable_rstudio_systemd_daemons' do
+    user "root"
+    ignore_failure true
+    code <<-EOF
       systemctl stop rstudio-server
       systemctl disable rstudio-server
     EOF
+  end
+
 end
